Move PriorityLevel to typing dir for circular imports
