+++ tests/sentry/issues/test_status_change_consumer.py
@@ -4,7 +4,6 @@ from typing import Any
 from unittest.mock import MagicMock, patch
 
 from sentry.issues.occurrence_consumer import _process_message
-from sentry.issues.priority import PriorityLevel
 from sentry.issues.status_change_consumer import bulk_get_groups_from_fingerprints
 from sentry.models.activity import Activity
 from sentry.models.group import Group, GroupStatus
@@ -12,7 +11,7 @@ from sentry.models.grouphistory import GroupHistory, GroupHistoryStatus
 from sentry.testutils.helpers.features import with_feature
 from sentry.testutils.pytest.fixtures import django_db_all
 from sentry.types.activity import ActivityType
-from sentry.types.group import GroupSubStatus
+from sentry.types.group import GroupSubStatus, PriorityLevel
 from tests.sentry.issues.test_occurrence_consumer import IssueOccurrenceTestBase, get_test_message
